api/requirements.txt
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.0
pydantic-settings>=2.0
structlog>=24.0
httpx>=0.27
solders>=0.21
anchorpy>=0.20
sse-starlette>=2.0
google-cloud-storage>=2.18
PyJWT>=2.9
